refactor: move progress bars to their domain packages
DisksProgressBar -> disks/cli/progress.py
CheckProgressBar -> preflight/progress.py
Delete clihelpers/progress.py (no longer needed)
