+++ nbkp/clihelpers/__init__.py
@@ -6,5 +6,3 @@ from .endpoints import (
     resolve_endpoints as resolve_endpoints,
 )
 from .output import OutputFormat as OutputFormat
-from .progress import CheckProgressBar as CheckProgressBar
-from .progress import DisksProgressBar as DisksProgressBar
